refactor: rename playtest-driver/types to interaction-driver/types (#516)
Completes the Phase 3 rename #515 deferred: playtest.ts and its own
utility module are gone, but scripts/shared/playtest-{driver,types}.ts
kept the old name even though they're now interaction-executor.ts's and
both interaction-mode runners' shared foundation, not playtest's own
code. Renamed to interaction-{driver,types}.ts, PlaytestFailure/
PlaytestGoal to InteractionFailure/InteractionGoal, and dropped the
exports nothing outside the module imports anymore (PlaytestBeat,
PlaytestDef, probe, tutorialState, describeAvailable, ProbedAction,
TutorialSnapshot) rather than carrying dead surface through the rename.

Updated every import site and repointed every live prose citation of
the old filenames across scripts/, src/, and tests/ so
NoDanglingDocReferences stays green. Two plan-doc passages and one
issue-479 doc citation predate the rename and describe design intent
at the time they were written, not current state — allowlisted with
the same historical-record rationale #515 already established for
similar entries, rather than rewritten out of the log.
