+++ scripts/shared/scenario-types.ts
@@ -42,8 +42,9 @@ export type InteractionStepAction =
   | { type: 'waitForSelector'; selector: string; timeout?: number }
   // Waits until the tutorial reaches one of the named steps (or ends), driving
   // the game's real auto-tick clock for the duration of the wait — the same
-  // bracketing the playtest harness uses — so queued work the step depends on
-  // (a surveyor walking out, a driver boarding) can actually finish. Fixed
+  // bracketing interaction-driver.ts's own awaitTutorialStep uses — so queued
+  // work the step depends on (a surveyor walking out, a driver boarding) can
+  // actually finish. Fixed
   // `wait` durations cannot express "the tutorial noticed", which is why
   // tutorial-driven scenarios desynced from the rails (#481). `timeout` is ms.
   | { type: 'waitForTutorialStep'; stepId: string | string[]; timeout?: number }
@@ -61,9 +62,9 @@ export type InteractionStepAction =
   // everything a player does, and the scenario vocabulary did not: there was
   // no way to set a <select>, so a step like `charge hole:* explosive:X
   // amount:Y` had no click-equivalent and stayed a console command. Each of
-  // these delegates to `runAction` in `playtest-driver.ts` — the same
-  // implementation the playtest channel uses, rather than a second copy that
-  // can drift from it.
+  // these delegates to `runAction` in `interaction-driver.ts` — the same
+  // implementation every interaction-mode caller uses, rather than a second
+  // copy that can drift from it.
   /** Set a form control's value the way typing or picking would. */
   | { type: 'set'; selector: string; value: string }
   /** Click the first usable control whose label matches (case-insensitive). */
@@ -151,12 +152,11 @@ export type ScenarioStepRole = 'player' | 'setup' | 'observe' | 'bootstrap' | 'g
 
 /**
  * What must be true after a step's actions ran. A scenario with no `expect`
- * anywhere only proves "nothing threw" — the playtest harness's whole
- * argument was that this is not the same as "the step actually happened"
- * (a click that silently no-ops still throws nothing). Field-for-field
- * mirror of `PlaytestGoal` (scripts/shared/playtest-types.ts) so interaction
- * mode reuses `checkGoal` from `playtest-driver.ts` directly — one evaluator,
- * not two that can drift apart.
+ * anywhere only proves "nothing threw", which is not the same as "the step
+ * actually happened" (a click that silently no-ops still throws nothing).
+ * Field-for-field mirror of `InteractionGoal` (scripts/shared/interaction-types.ts)
+ * so interaction mode reuses `checkGoal` from `interaction-driver.ts` directly
+ * — one evaluator, not two that can drift apart.
  *
  * `usable`/`blocked`/`tutorialStep` need a live page and are only checked
  * when the scenario runs in interaction mode; command mode has no DOM, so it
