+++ tests/unit/scenario-defs.test.ts
@@ -881,11 +881,11 @@ describe('Role-marked steps obey checkStepActionAllowed for bootstrap/guard (iss
 // ──────────────────────────────────────────────
 // 15. A step's `expect`, when present, is shaped correctly and checkable
 // (issue #479 follow-up: scenarios gained assertions instead of proving only
-// "the command didn't throw" — mirrors playtest-defs.test.ts's equivalent
-// rule for beats). Checked in BOTH modes: command mode via
+// "the command didn't throw" — mirrored the now-deleted playtest-defs.test.ts's
+// equivalent rule for beats). Checked in BOTH modes: command mode via
 // checkGoalAgainstState (equals/increased only — no DOM), interaction mode
 // via checkGoal (all fields) — scripts/shared/scenario-goal.ts and
-// scripts/shared/playtest-driver.ts respectively.
+// scripts/shared/interaction-driver.ts respectively.
 // ──────────────────────────────────────────────
 describe('Step expect field is shaped correctly when present', () => {
   for (const name of ALL_SCENARIO_NAMES) {
